[feat] #13 재사용 전역 타입 정의
