+++ src/@types/HttpFailed.d.ts
@@ -1,0 +1,7 @@
+declare module "http" {
+  interface HttpFailed {
+    statusCode: number;
+    message: string;
+    error?: string;
+  }
+}
