+++ src/@types/todo.d.ts
@@ -1,0 +1,8 @@
+declare module "todo" {
+  interface Todo {
+    id: number;
+    todo: string;
+    isCompleted: boolean;
+    userId: number;
+  }
+}
