+++ src/@types/user.d.ts
@@ -1,0 +1,6 @@
+declare module "user" {
+  interface User {
+    email: string;
+    password: string;
+  }
+}
